fix book list layout
