+++ src/app/book-list/components/BookGrid.tsx
@@ -2,7 +2,39 @@ import { Box } from '@mui/material';
 import BookItem from '@/app/book-list/components/BookItem';
 import React from 'react';
 import { Book } from '@/types/Book';
-import { dummyBooks } from '@/app/book-list/components/BookGrid.stories';
+
+export const dummyBooks: Book[] = [
+  {
+    id: 1,
+    title: 'Book1',
+    status: 'unread',
+    author: 'Author1',
+  },
+  {
+    id: 2,
+    title: 'Book2',
+    status: 'unread',
+    author: 'Author2',
+  },
+  {
+    id: 3,
+    title: 'Book3',
+    imgUrl: 'https://placehold.jp/150x150.png',
+    status: 'unread',
+    author: 'Author3',
+  },
+  {
+    id: 4,
+    title: 'Book4',
+    status: 'reading',
+    imgUrl: 'https://placehold.jp/240x160.png',
+  },
+  {
+    id: 5,
+    title: 'Book5',
+    status: 'done',
+  },
+];
 
 interface BookGridProps {
   books: Book[];
