+++ src/app/book-list/page.tsx
@@ -1,27 +1,46 @@
 import React from 'react';
-import BookGrid from '@/app/book-list/components/BookGrid';
-import { dummyBooks } from '@/app/book-list/components/BookGrid.stories';
+import BookGrid, { dummyBooks } from '@/app/book-list/components/BookGrid';
+import { Box, Button } from '@mui/material';
+import BasicTabs from '@/components/BasicTabs';
 
-// function StatusTabs() {
-//   const [status, setStatus] = React.useState('unread');
-//
-//   const handleStatusChange = (newStatus: string) => {
-//     setStatus(newStatus);
-//   };
-//
-//   return (
-//     <Box sx={{ width: '100%' }}>
-//       <Box>Unread</Box>
-//       <Box>Reading</Box>
-//       <Box>Done</Box>
-//     </Box>
-//   );
-// }
+const readingStatusOptions = [
+  { label: 'all' },
+  { label: 'unread' },
+  { label: 'reading' },
+  { label: 'done' },
+];
 
 export default function BookList() {
   return (
-    <div>
-      <BookGrid books={dummyBooks} />
-    </div>
+    <Box sx={{ width: '100%' }}>
+      <Box
+        sx={{
+          display: 'flex',
+          justifyContent: 'space-between',
+          alignItems: 'center',
+          borderBottom: 1,
+          borderColor: 'divider',
+        }}
+      >
+        <BasicTabs
+          onChange={(v) => {
+            console.log('Selected value:', v);
+          }}
+          options={readingStatusOptions}
+        />
+        <Button
+          variant={'contained'}
+          color={'primary'}
+          onClick={() => {
+            console.log('onClick register button');
+          }}
+        >
+          新規登録
+        </Button>
+      </Box>
+      <div>
+        <BookGrid books={dummyBooks} />
+      </div>
+    </Box>
   );
 }
