+++ src/components/BasicTabs.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
-import Box from '@mui/material/Box';
 
 interface Option {
   label: string;
@@ -21,14 +20,10 @@ export default function BasicTabs({ options, onChange }: BasicTabsProp) {
   };
 
   return (
-    <Box sx={{ width: '100%' }}>
-      <Box>
-        <Tabs value={value} onChange={(e, v) => handleChange(e, v)}>
-          {options.map((option) => (
-            <Tab key={option.label} label={option.label} />
-          ))}
-        </Tabs>
-      </Box>
-    </Box>
+    <Tabs value={value} onChange={(e, v) => handleChange(e, v)}>
+      {options.map((option) => (
+        <Tab key={option.label} label={option.label} />
+      ))}
+    </Tabs>
   );
 }
